pyproject.toml
[project]
name = "crupto-curr-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai-agents>=0.0.19",
    "python-dotenv>=1.1.1",
    "streamlit>=1.46.0",
]
